split probe
